refactor: cryptoprovider types
